requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
openai==0.27.10
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.0.4
yarl==1.9.2
